backend/requirements.txt
fastapi
uvicorn[standard]
pymongo
python-dotenv
python-multipart
httpx
pydantic
passlib
python-jose[cryptography]
cloudinary
google-auth
requests